Concept Notes and Specifications
EMPACT Africa is about to launch an online initiative to help local
faith communities fight the stigma of HIV and AIDS.  This PowerPoint
slide deck includes descriptions of the vision and key concepts, as
well as usage scenarios and a draft data model.  We are looking for
implementation and deployment partners for the initiative.